Fixes a bug in progress reporting in TextFileReader
It was 1) thread-unsafe (accessed a reader's variable from a different
thread) and 2) incorrect (did not return position of the last returned
record).

Because of this, the reader could end up in a situation where it would
keep returning a position past the end of the range, confusing dynamic
work rebalancing.

----Release Notes----
[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=119983487