Use shelve as a disk backed dictionary optionally in PValueCache
A new DirectRunner based DiskCachedPipelineRunner is introduced. Shelve
will automatically spill dictionary entries into disk, reducing the
memory requirement. For small pipelines the performance impact is
minimal as it has an in memory cache for recent objects. For large
pipelines that requires multiple pcollection object to be in the cache
at the same time it prevents OOMs. However it will have a performance
impact for large pipelines because of disk IO.

Memory requirement of this new runner is capped by the single
ptransform in the pipeline that consumes the largest total input
(input + side inputs in aggregate).

----Release Notes----
[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=123441597