Reintroduce OperationCounters.should_sample
This change was previously reverted for a performance regression.  That
regression was thought to be due to should_sample() being slow, but that is not
the case; rather, the previous partial implementation of do_sample() was
causing the slowdown.

----Release Notes----
[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=124573546